вынесла хук useFullscreen из FullscreenButton
