+++ packages/client/src/shared/lib/fullscreen/index.ts
@@ -1,1 +1,2 @@
+export * from './useFullscreen';
 export * from './fullscreenUtils';
